backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis for caching and sessions
redis==5.0.1
celery==5.3.4

# Google Cloud and NotebookLM
google-cloud-storage==2.10.0
google-cloud-secretmanager==2.17.0
google-auth==2.23.4
google-api-python-client==2.108.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.0.3

# HTTP client
httpx==0.25.2
requests==2.31.0

# File processing
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2

# Logging and monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Security
python-decouple==3.8
cryptography==41.0.7

# Utilities
python-slugify==8.0.1
email-validator==2.1.0
jinja2==3.1.2
